requirements.txt
grpcio
grpcio-tools
torch
torchvision
ultralytics